:recycle: refactor: improving the legebility for params arguments  and create type interface for this
