+++ src/types/types.ts
@@ -1,0 +1,5 @@
+export interface CreateUserInput {
+    name:string,
+    email: string,
+    description: string | null
+}
