+++ tsconfig.json
@@ -8,7 +8,7 @@
      "moduleResolution": "NodeNext",
      "skipLibCheck": true,
      "forceConsistentCasingInFileNames": true,
-     "noEmitOnError": false,
+     "noEmitOnError": true,
      "outDir": "./dist",
      "rootDir": "./src"
    },
